pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
]
